Improve speed management

+++ src/entity/tile/Brick.java
@@ -4,11 +4,11 @@ import graphic.Sprite;
 
 public class Brick extends Tile {
 
-    public Brick(int x, int y) {
+    public Brick(double x, double y) {
         super(x, y);
     }
 
-    public Brick(int x, int y, Sprite[] sprite) {
+    public Brick(double x, double y, Sprite[] sprite) {
         super(x, y, sprite);
     }
